Move Read and BibleVerses models
