+++ common/lessons-data/src/main/java/app/ss/lessons/data/model/SSRead.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 Adventech <info@adventech.io>
+ * Copyright (c) 2021. Adventech <info@adventech.io>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cryart.sabbathschool.reader.data.model
+package app.ss.lessons.data.model
 
 import androidx.annotation.Keep
 import com.google.firebase.database.DataSnapshot
