+++ features/bible/build.gradle.kts
@@ -62,6 +62,7 @@ dependencies {
     implementation(project(BuildModules.Common.CORE))
     implementation(project(BuildModules.Common.DESIGN))
     implementation(project(BuildModules.Common.TRANSLATIONS))
+    implementation(project(BuildModules.Common.LESSONS_DATA))
     implementation(project(BuildModules.Features.READER))
 
     implementation(Kotlin.KOTLIN)
